requirements.txt
jupyter
opencv-contrib-python
matplotlib
scipy
Pillow